+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased - 1.0.5
 
 - Shorten the pubspec.yaml description's length to increase the pub score.
+- Add animation and adjust styling on LitLockIcon.
+
 ## 1.0.4
 
 - Remove `CustomInfoBar`, `DummyInfoBar` and `TransparentInfoBar`. Please use the existing `LitSnackbar` widgets to display feedback messages.